+++ HW07/task1_thrust.cu
@@ -21,7 +21,7 @@ void generate(T *output, std::size_t n, float lower, float upper) {
 
 
 int main(int argc, char *argv[]) {
-  const unsigned int n = atoi(argv[1]);
+  const unsigned int n = atol(argv[1]);
 
   float *data = new float[n];
   generate<float>(data, n, -1.0, 1.0);
